Vueで問い合わせ内容を保存
