+++ frontend/package.json
@@ -11,7 +11,8 @@
   "dependencies": {
     "axios": "^1.6.8",
     "pinia": "^2.1.7",
-    "vue": "^3.4.21"
+    "vue": "^3.4.21",
+    "vue-router": "^4.3.2"
   },
   "devDependencies": {
     "@vitejs/plugin-vue": "^5.0.4",
