+++ frontend/src/main.ts
@@ -1,5 +1,6 @@
 import { createPinia } from 'pinia'
 import { createApp } from 'vue'
+import { router } from "./router";
 
 import App from './App.vue'
 import './style.css'
@@ -7,4 +8,5 @@ import './style.css'
 const app = createApp(App);
 
 app.use(createPinia());
+app.use(router);
 app.mount('#app');
